新增部分功能

+++ Pages/SendAndReceiveJpeg.xaml.cs
@@ -24,11 +24,15 @@ namespace SQClient.Pages
     /// </summary>
     public partial class SendAndReceiveJpeg : BaseUart
     {
-        const string strFilter = "JPEG files (*.jpeg *.jpg)|*.jpeg;*.jpg";
-        bool isInReceive = false;
-        int receiveLen;
+        const string strFilter = "JPEG files (*.jpg *.jpeg)|*.jpg;*.jpeg|All files (*.*)|*.*";
+        const uint RGB565Length = 320 * 240 * 2;
+
+        bool? isInReceive = false, isReceiveRgb565 = false;
+        uint receiveLen;
         DispatcherTimer timer;
         DateTime startTime;
+        string rgb565path;
+
 
         public SendAndReceiveJpeg()
         {
@@ -56,7 +60,16 @@ namespace SQClient.Pages
                 if (this.procbar.Value == this.procbar.Maximum)
                 {
                     StopReceive();
-                    this.img.Source = Tools.ByteArrayToBitmapImage(ListRead.ToArray());                    
+
+                    if (isInReceive == null)
+                    {
+                        this.img.Source = Tools.ByteArrayToBitmapImage(ListRead.ToArray());
+                    }
+                    else
+                    {
+                        System.Drawing.Bitmap bitmap = Tools.BytesRGB565ToBitmap(320, 240, ListRead.ToArray());
+                        this.img.Source = Tools.BitmapToBitmapSource(bitmap);
+                    }                    
                 }
             };
          }
@@ -86,16 +99,14 @@ namespace SQClient.Pages
                 }
 
                 this.tglbtnConnect.Content = "断开";
-                this.btnReceive.IsEnabled = true;
-                this.btnSend.IsEnabled = true;
+                this.btnReceive.IsEnabled = true;                
             }
             else //断开
             {
                 Disconnect();
 
                 this.tglbtnConnect.Content = "连接";
-                this.btnReceive.IsEnabled = false;
-                this.btnSend.IsEnabled = false;
+                this.btnReceive.IsEnabled = false;                
             }
         }
 
@@ -147,21 +158,35 @@ namespace SQClient.Pages
         void StopReceive()
         {
             timer.Stop();
-            this.btnSaveAs.Dispatcher.Invoke(() => this.btnSaveAs.Visibility = Visibility.Visible);
+
+            if (isInReceive == null)
+            {
+                this.btnSaveAs.Dispatcher.Invoke(() => this.btnSaveAs.Visibility = Visibility.Visible);
+            }
         }
 
         protected override void OnDataReceived()
         {
             base.OnDataReceived();
 
-            if (!isInReceive && ListRead.Count > 7)
+            if (isReceiveRgb565.Value)
+            {
+                if (ListRead.Count == RGB565Length)
+                {
+                    File.WriteAllBytes(rgb565path, ListRead.ToArray());
+                    isReceiveRgb565 = null;                    
+                }
+                return;
+            }
+
+            if (isInReceive != null && !isInReceive.Value && ListRead.Count > 7)
             {
                 for (int i = 0; i < ListRead.Count; i++)
                 {
                     if (i + 7 < ListRead.Count && ListRead[i] == 0xA5 && ListRead[i+1] == 0x5A && ListRead[i+2] == 'S')
                     {
-                        receiveLen = (ListRead[i+3] << 24) | (ListRead[i+4] << 16) | (ListRead[i+5] << 8) | (ListRead[i+6]);    
-                        isInReceive = true;
+                        receiveLen = (uint)((ListRead[i+3] << 24) | (ListRead[i+4] << 16) | (ListRead[i+5] << 8) | (ListRead[i+6]));    
+                        isInReceive = null;
                         ListRead.RemoveRange(0, i + 8); //最后一位为校验位
                         StartReceive();                                            
                         break;
@@ -192,9 +217,52 @@ namespace SQClient.Pages
             }
         }
 
-        private void CameraOnClick(object sender, RoutedEventArgs e)
+        private void RegisterOnClick(object sender, RoutedEventArgs e)
         {
+            //Write(TestCmds.CMD_REGISTER());
+            
+            OpenFileDialog dlg = new OpenFileDialog();
 
+            if ((bool)dlg.ShowDialog(Application.Current.MainWindow))
+            {               
+                Write(TestCmds.CMD_REGISTER());
+
+                //TODO: 发送rbg565流
+                //FileStream stream = new FileStream(dlg.FileName, FileMode.Open);
+
+                //byte[] readBuf = new byte[stream.Length];
+                //stream.Read(readBuf, 0, readBuf.Length);
+
+                //Write(readBuf);
+                //stream.Close();
+
+                byte[] buf = RGB565.GetRGB565Image(dlg.FileName);
+                Write(buf);
+            }
+        }
+
+        private void RGB565OnClick(object sender, RoutedEventArgs e)
+        {
+            SaveFileDialog dlg = new SaveFileDialog();
+            dlg.Filter = "txt file|*.txt";
+
+            if ((bool)dlg.ShowDialog(Application.Current.MainWindow))
+            {
+                rgb565path = dlg.FileName;
+                ListRead.Clear();
+                isReceiveRgb565 = true;
+                StartReceive();
+                receiveLen = RGB565Length;
+
+                if ((sender as Button).Tag.ToString() == "JPEG")
+                {
+                    Write(TestCmds.CMD_JPEG_RGB565());
+                }
+                else
+                {
+                    Write(TestCmds.CMD_RGB565());
+                }
+            }
         }
     }
 }
